backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyJWT[crypto]==2.8.0
cryptography==41.0.7
asyncpg==0.29.0
redis[hiredis]==5.0.1
pandas==2.1.3
# pandas-ta  # Temporarily disabled - package not available on PyPI for Python 3.11
numpy==1.26.2
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
python-json-logger==2.0.7
httpx==0.25.2
websockets==12.0
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
locust==2.20.0
python-dotenv==1.0.0